requirements.txt
discord==0.16.12
python-dotenv==0.10.1
pymongo==3.7.2
dateutil==2.8.0